Refactored unet model.
