+++ src/score_models/train_steps/ddpm_train_step.py
@@ -1,4 +1,4 @@
-from typing import List
+from typing import List, Optional
 
 import torch
 import torch.nn as nn
@@ -16,7 +16,7 @@ class TrainStepDDPM(TrainStep):
         # define loss functions
         self.loss_fn = DenoiseLoss()
 
-    def __call__(self, x: Tensor) -> Tensor:
+    def __call__(self, x: Tensor, y: Optional[Tensor] = None) -> Tensor:
         """Computes the loss for the Denoising Score Matching.
 
         :param x: batch of input tensors
@@ -38,7 +38,7 @@ class TrainStepDDPM(TrainStep):
         # from noisy example
         eps = torch.randn_like(x)
         x_tilde = torch.sqrt(alphas_bar) * x + torch.sqrt(1 - alphas_bar) * eps
-        eps_pred = self.score_model(x_tilde, indices)
+        eps_pred = self.score_model(x_tilde, y, indices) if y is not None else self.score_model(x_tilde, indices)
 
         # compute and return loss
         return self.loss_fn(eps_pred=eps_pred, eps=eps)
